Move `fetch_atoms` to `quacc.schemas` (#844)
