+++ src/quacc/recipes/emt/core.py
@@ -11,10 +11,10 @@ from ase.calculators.emt import EMT
 from ase.optimize import FIRE
 
 from quacc import job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.utils.calc import run_ase_opt, run_calc
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
