+++ src/quacc/recipes/emt/defects.py
@@ -8,8 +8,8 @@ from pymatgen.analysis.defects.generators import VacancyGenerator
 from quacc import flow, job, subflow
 from quacc.recipes.emt.core import relax_job as _relax_job
 from quacc.recipes.emt.core import static_job as _static_job
+from quacc.schemas import fetch_atoms
 from quacc.utils.defects import make_defects_from_bulk
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
