+++ src/quacc/recipes/emt/slabs.py
@@ -6,8 +6,8 @@ from typing import TYPE_CHECKING
 from quacc import flow, job, subflow
 from quacc.recipes.emt.core import relax_job as _relax_job
 from quacc.recipes.emt.core import static_job as _static_job
+from quacc.schemas import fetch_atoms
 from quacc.utils.slabs import make_max_slabs_from_bulk
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
