+++ src/quacc/recipes/gulp/core.py
@@ -7,10 +7,10 @@ from typing import TYPE_CHECKING
 from ase.calculators.gulp import GULP
 
 from quacc import job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_run
 from quacc.utils.calc import run_calc
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
