+++ src/quacc/recipes/lj/core.py
@@ -11,6 +11,7 @@ from ase.calculators.lj import LennardJones
 from ase.optimize import FIRE
 
 from quacc import job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import (
     summarize_opt_run,
     summarize_run,
@@ -20,7 +21,6 @@ from quacc.schemas.ase import (
 from quacc.utils.calc import run_ase_opt, run_ase_vib, run_calc
 from quacc.utils.dicts import merge_dicts
 from quacc.utils.thermo import ideal_gas
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
