+++ src/quacc/recipes/newtonnet/core.py
@@ -10,6 +10,7 @@ from ase.vibrations.data import VibrationsData
 from monty.dev import requires
 
 from quacc import SETTINGS, job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import (
     summarize_opt_run,
     summarize_run,
@@ -19,7 +20,6 @@ from quacc.schemas.ase import (
 from quacc.utils.calc import run_ase_opt, run_calc
 from quacc.utils.dicts import merge_dicts
 from quacc.utils.thermo import ideal_gas
-from quacc.utils.wflows import fetch_atoms
 
 try:
     from sella import Sella
