+++ src/quacc/recipes/newtonnet/ts.py
@@ -12,10 +12,10 @@ from quacc import SETTINGS, job
 from quacc.recipes.newtonnet.core import _add_stdev_and_hess
 from quacc.recipes.newtonnet.core import freq_job as _freq_job
 from quacc.recipes.newtonnet.core import relax_job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_opt_run
 from quacc.utils.calc import run_ase_opt
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.wflows import fetch_atoms
 
 try:
     from sella import IRC, Sella
