+++ src/quacc/recipes/psi4/core.py
@@ -7,11 +7,11 @@ from ase.calculators.psi4 import Psi4
 from monty.dev import requires
 
 from quacc import job
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_run
 from quacc.utils.atoms import get_charge, get_multiplicity
 from quacc.utils.calc import run_calc
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.wflows import fetch_atoms
 
 try:
     import psi4
