+++ src/quacc/recipes/qchem/core.py
@@ -8,11 +8,11 @@ from ase.optimize import FIRE
 
 from quacc import job
 from quacc.calculators.qchem import QChem
+from quacc.schemas import fetch_atoms
 from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.utils.atoms import check_charge_and_spin
 from quacc.utils.calc import run_ase_opt, run_calc
 from quacc.utils.dicts import merge_dicts, remove_dict_empties
-from quacc.utils.wflows import fetch_atoms
 
 try:
     from sella import Sella
