+++ src/quacc/recipes/vasp/core.py
@@ -5,10 +5,10 @@ from typing import TYPE_CHECKING
 
 from quacc import job
 from quacc.calculators.vasp import Vasp
+from quacc.schemas import fetch_atoms
 from quacc.schemas.vasp import summarize_run
 from quacc.utils.calc import run_calc
 from quacc.utils.dicts import merge_dicts
-from quacc.utils.wflows import fetch_atoms
 
 if TYPE_CHECKING:
     from ase import Atoms
