+++ src/quacc/schemas/__init__.py
@@ -1,1 +1,5 @@
 """Schemas for storing data from recipes"""
+
+from quacc.schemas.atoms import fetch_atoms
+
+__all__ = ["fetch_atoms"]
